+++ grovedb/src/debugger.rs
@@ -911,13 +911,17 @@ fn element_to_grovedbg(element: crate::Element) -> grovedbg_types::Element {
                 element_flags,
             }
         }
-        crate::Element::ItemWithSumItem(value, sum_value, element_flags) => {
-            grovedbg_types::Element::ItemWithSumItem {
-                value,
-                sum_item_value: sum_value,
-                element_flags,
-            }
-        }
+        crate::Element::ItemWithSumItem(value, sum_value, element_flags)
+        | crate::Element::ItemWithSumItemWithBackwardsReferences(
+            value,
+            sum_value,
+            _,
+            element_flags,
+        ) => grovedbg_types::Element::ItemWithSumItem {
+            value,
+            sum_item_value: sum_value,
+            element_flags,
+        },
         crate::Element::SumTree(root_key, sum, element_flags) => grovedbg_types::Element::Sumtree {
             root_key,
             sum,