+++ grovedb/src/element/aggregate_sum_query/mod.rs
@@ -807,7 +807,8 @@ impl ElementAggregateSumQueryExtensions for Element {
         let value = match element.into_underlying() {
             Element::SumItem(value, _)
             | Element::ItemWithSumItem(_, value, _)
-            | Element::SumItemWithBackwardsReferences(value, _, _) => value,
+            | Element::SumItemWithBackwardsReferences(value, _, _)
+            | Element::ItemWithSumItemWithBackwardsReferences(_, value, _, _) => value,
             _ => return Err(Error::InvalidInput("Only sum items are allowed")),
         };