+++ grovedb/src/operations/delete/delete_internal_on_transaction/v2.rs
@@ -363,6 +363,7 @@ impl<'b, B: AsRef<[u8]>> Visit<'b, B> for DeletionVisitor<'_, '_, 'b, B> {
                 element,
                 Element::ItemWithBackwardsReferences(..)
                     | Element::SumItemWithBackwardsReferences(..)
+                    | Element::ItemWithSumItemWithBackwardsReferences(..)
                     | Element::BidirectionalReference(..)
             )
         {