+++ grovedb/src/operations/proof/verify.rs
@@ -1259,7 +1259,8 @@ impl GroveDb {
             let value = match element.into_underlying() {
                 Element::SumItem(value, _)
                 | Element::ItemWithSumItem(_, value, _)
-                | Element::SumItemWithBackwardsReferences(value, _, _) => value,
+                | Element::SumItemWithBackwardsReferences(value, _, _)
+                | Element::ItemWithSumItemWithBackwardsReferences(_, value, _, _) => value,
                 _ => {
                     return Err(Error::InvalidProof(
                         query.clone(),
@@ -2081,7 +2082,8 @@ impl GroveDb {
                             | Element::ReferenceWithSumItem(..)
                             | Element::BidirectionalReference(..)
                             | Element::ItemWithBackwardsReferences(..)
-                            | Element::SumItemWithBackwardsReferences(..) => {
+                            | Element::SumItemWithBackwardsReferences(..)
+                            | Element::ItemWithSumItemWithBackwardsReferences(..) => {
                                 return Err(Error::InvalidProof(
                                     query.clone(),
                                     "V1 proof has lower layer for a non-tree element.".to_string(),
@@ -3261,7 +3263,8 @@ impl GroveDb {
                             | Element::ReferenceWithSumItem(..)
                             | Element::BidirectionalReference(..)
                             | Element::ItemWithBackwardsReferences(..)
-                            | Element::SumItemWithBackwardsReferences(..) => {
+                            | Element::SumItemWithBackwardsReferences(..)
+                            | Element::ItemWithSumItemWithBackwardsReferences(..) => {
                                 return Err(Error::InvalidProof(
                                     query.clone(),
                                     "Proof has lower layer for a non Tree.".to_string(),